JessamineCLI

Documentation for JessamineCLI.

JessamineCLI.EpochSpec

JessamineCLI.ExploreSimplifySearchSpec

JessamineCLI.conf_convert

JessamineCLI.get_or_parse

JessamineCLI.mkpath_and_open

JessamineCLI.mkpath_and_open

JessamineCLI.parse_op_weight_scheme

JessamineCLI.split_on_semicolons

JessamineCLI.split_symbols

JessamineCLI.zippairs

JessamineCLI.@cfield

JessamineCLI.EpochSpec — Type

EpocSpec

Configuration for one evolution epoch

JessamineCLI.ExploreSimplifySearchSpec — Type

ExploreSimplifySearchSpec

Specifies a search process in which the exploration EvolutionSpec is followed to form the initial population and to run the bulk of the evolutionary search. Then a simplification EvolutionSpec is used if given.

JessamineCLI.conf_convert — Method

conf_convert(type, value)

Use convert or pyconvert to convert value to type.

JessamineCLI.get_or_parse — Function

get_or_parse(spec, key, default_value, etype=typeof(default_value))

Try to get the value corresponding to key from the dictionary spec. If the result is nothing, return that. If the result is a string and etype is not a string type, parse the value into the specified type. If the result is some other type, convert the value to the specified type.

JessamineCLI.mkpath_and_open — Method

mkpath_and_open(f, path, args...; kwargs...)

Ensure that the directory part of path exists, then call open on f and the whole path, splatting in args and kwargs.

JessamineCLI.mkpath_and_open — Method

mkpath_and_open(path, args...; kwargs...)

Ensure that the directory part of path exists, then call open on the whole path, splatting in args and kwargs.

JessamineCLI.parse_op_weight_scheme — Function

parse_op_weight_scheme(ps)

Parse a StandardWeightScheme

JessamineCLI.split_on_semicolons — Method

split_on_semicolons(s)

Split a string into substrings on semicolons, stripping out loose whitespace.

JessamineCLI.split_symbols — Method

split_symbols(s)

Given a string, pull out the sequences of characters that are not whitespace, a comma, a bracket, a brace, or a quotation mark. Return themn as an array. This effectively splits a string of space- or comma-delimited symbols (identifiers), with our without quotation marks, into just the symbols.

JessamineCLI.zippairs — Method

zippairs(a, b)

Return an iterator that produces [a[1] => b[1], a[2] => b[2], ...]

JessamineCLI.@cfield — Macro

@cfield spec key default_value etype

Extract a configuration field from spec (a Dict or similar) using key as the lookup key, and default_value if it's not present. Strings are be parsed using parse and eval. Results are converted to etype, which defaults to the type of default_value. Produce syntax of the form

key = ...

suitable for inclusion as a variable assignment, field in a named tuple, keyword argument to a function call, etc.

Note that for inclusion in a named tuple or function call, the @cfield spec key value has to be on a line by itself, so the comma has to go on the next line.

If etype is omitted, it defaults to the type of default_value.